--- src/app/api/err/browser/everytime/route.js
@@ -1,5 +1,0 @@
-import { handleApiRequest } from '@/app/api/apiWrapper';
-
-export async function GET(request, { params }) {
-  return handleApiRequest('GET', '/err/browser/everytime', { params });
-}
