--- src/app/api/err/browser/route.js
@@ -1,5 +1,0 @@
-import { handleApiRequest } from '@/app/api/apiWrapper';
-
-export async function GET(request, { params }) {
-  return handleApiRequest('GET', '/err/browser', { params });
-}
