+++ src/app/checkInAppBrowser.jsx
@@ -2,8 +2,8 @@
 
 import './page.css';
 import { useEffect, useState } from 'react';
-import Image from 'next/image';
 import { MainLogoImage } from '@/components/common/MainLogoImage';
+import InAppBrowserOutButton from '@components/InAppBrowserOutButton';
 
 const IN_APP_BROWSER_NAMES = {
   kakaotalk: '카카오톡',
@@ -11,54 +11,6 @@ const IN_APP_BROWSER_NAMES = {
   instagram: '인스타그램',
   line: '라인',
 };
-
-function InAppBrowserOutButton() {
-  const [isRedirectPossible, setIsRedirectPossible] = useState(false);
-  const useragt = navigator.userAgent.toLowerCase();
-
-  useEffect(() => {
-    if (useragt.match(/kakaotalk|line/i)) {
-      setIsRedirectPossible(true);
-    }
-  }, [useragt]);
-
-  const onClick = () => {
-    const target_url = window.location.href;
-    if (useragt.match('kakaotalk')) {
-      window.location.href =
-        'kakaotalk://web/openExternal?url=' + encodeURIComponent(target_url);
-    } else if (useragt.match('line')) {
-      if (target_url.indexOf('?') !== -1) {
-        window.location.href = target_url + '&openExternalBrowser=1';
-      } else {
-        window.location.href = target_url + '?openExternalBrowser=1';
-      }
-    }
-  };
-
-  if (!isRedirectPossible) {
-    if (!useragt.match('everytimeapp')) return;
-    if (useragt.match(/ios|ipad|ipod/i)) {
-      return (
-        <p>
-          상단의 화살표 아이콘({' '}
-          <Image
-            src="/vectors/open-external-link-icon.svg"
-            width={12}
-            height={12}
-            className="ico"
-            alt="arrow button"
-          />{' '}
-          ) 클릭 &rarr; 외부 브라우저로 열기
-        </p>
-      );
-    } else {
-      return <p>상단의 점 세 개( ⋮ ) 클릭 &rarr; 외부 브라우저로 열기</p>;
-    }
-  }
-
-  return <button onClick={onClick}>외부 브라우저로 이동</button>;
-}
 
 export const MainLogo = () => {
   const [isInAppBrowser, setIsInAppBrowser] = useState(false);
